packed caps

+++ test/integration/BlueIntegrationTest.sol
@@ -104,11 +104,11 @@ contract BlueIntegrationTest is BaseTest {
 
     function increaseAbsoluteAndRelativeCapToMax(bytes memory idData) internal {
         vm.startPrank(curator);
-        vault.submit(abi.encodeCall(IVaultV2.increaseAbsoluteCap, (idData, type(uint256).max)));
+        vault.submit(abi.encodeCall(IVaultV2.increaseAbsoluteCap, (idData, type(uint128).max)));
         vault.submit(abi.encodeCall(IVaultV2.increaseRelativeCap, (idData, WAD)));
         vm.stopPrank();
 
-        vault.increaseAbsoluteCap(idData, type(uint256).max);
+        vault.increaseAbsoluteCap(idData, type(uint128).max);
         vault.increaseRelativeCap(idData, WAD);
     }
 }
